services/auth/requirements.txt
flask
python-dotenv
PyJWT
pymongo
